% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslc.R
\name{get.brp.taxa}
\alias{get.brp.taxa}
\title{Run The Two-Stagewise Method for Differential Abundance Analysis}
\usage{
get.brp.taxa(wmgs.dat, meta.dat, f.cutoff = 0.1, tax.cutoff = 0.1)
}
\arguments{
\item{wmgs.dat}{A data frame containing information on the functions and taxa associated with each function for all samples.
The first column should be labeled X..Pathway and contain the names of the functions and taxa linked with each function. The
remaining columns should contain the sample IDs and list the counts for each function and linked taxon.}

\item{meta.dat}{A data frame containing information on the outcome variable and any covariates for each subject. This must have one
column labeled "SampleID" to denote the participants and another column labeled "y" to denote the outcome variable. The remaining
columns should include information on any covariates of interest.}

\item{f.cutoff}{Adjusted p-value threshold to use for the first stage to identify DA functions.}

\item{tax.cutoff}{Adjusted p-value threshold to use for the second stage to identify BRp taxa.}
}
\value{
This function returns a list containing the following:

fun.res: A data frame containing the results from running the first stage of the method where we identify DA functions.
tax.res: A list containing the results from the second stage where we try to identify BRp taxa given each function. The names
of each element of the list correspond to the functions, and the data frame contains the results from the second stage.
brp.tax: A vector containing the names of all identified BRp taxa.
}
\description{
This is a function that takes a data frame containing the abundances of the functions along with their associated taxa and
a data frame containing the outcome variable and any covariates of interest. It runs the two-stagewise approach to finding
BRp taxa and returns the results from both stages along with a list of all identified BRp taxa.
}
