% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twostage.R
\name{get.taxa}
\alias{get.taxa}
\title{Obtain Taxa Associated With Each Pathway}
\usage{
get.taxa(func)
}
\arguments{
\item{func}{A data frame where the first column contains the pathways and associated taxa and the remaining
columns contain the sample IDs. The first column must have the name X..Pathway.}
}
\value{
This function returns a list where the names are the gene pathways and each element of the list
is a vector containing all species associated with the corresponding pathway.
}
\description{
This is a function that takes the output from Humann as an input and returns a list of all
the species associated with each of the pathways.
}
