% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Twostage.R
\name{two.stage.da.analysis}
\alias{two.stage.da.analysis}
\title{Run The Two-Stagewise Method for Differential Abundance Analysis}
\usage{
two.stage.da.analysis(y, fun_prof, tax_prof, Fjs, x, cutoff = 0.1)
}
\arguments{
\item{y}{A vector with elements from {0, 1} to indicate which population each sample belongs to. The order
of the elements of y must match the order of the samples in the rows of the functional profile, taxonomic
profile, and matrix of covariates.}

\item{fun_prof}{A matrix with the samples as rows and pathways as columns. Note that the abundances of
each sample must be converted to proportions in order for this to work. Also, the order of the rows
must match the order of the samples in the response vector.}

\item{tax_prof}{A matrix with the samples as rows and taxa as columns. Note that the abundances of each sample
must be converted to proportions in order for this to work. The rownames of the taxonomic profile must
match up to the rownames of the functional profile.}

\item{Fjs}{A list where the names of the list are the same as the column names of the functional profile,
and each element of the list is a vector indicating which microbes in the taxonomic profile are associated
with the corresponding function. The elements of the vector can be given as either column indices or names.}

\item{x}{A matrix where the samples are rows and the columns are covariates to be adjusted for in the model.
The rownames must match up to those of the functional and taxonomic profile, and any NA values must be
removed from this matrix before running the method.}

\item{cutoff}{A number to indicate which threshold the adjusted p-value should be cutoff at. All taxa with
an adjusted p-value below this threshold will be detected as differentially abundant. To correct for
multiple comparisons, we use the Benjamini-Hochberg method.}
}
\value{
This function returns a list with the following elements.

da.fun: The names of all differentially abundant pathways.

da.taxa: The names of all differentially abundant taxa conditional on differentially abundant pathways.

result: A data frame where the columns are each taxa name, their nominal p-value, and adjusted p-value.
}
\description{
This is a function that takes a vector indicating the group membership, the functional profile,
the taxonomic profile, a list containing all taxa associated with each function, and a matrix of covariates
and runs the two-stagewise approach to finding differentially abundant taxa given differentially
abundant functions.
}
